SkylandBeeMutation.java
I propose to change Bee breeding mutation in normal apiary for the use of a modpack focused on bees and starting in void with only bees production, so i need a BeeMutation.class that require to eliminate block requirements under apiary to have mutation of the bees. So if you could change or create a patch to do this could be really good.